+++ cmd/auth/constants.go
@@ -1,8 +1,0 @@
-package auth
-
-const (
-	codeChallengeLength = 87
-	supabaseID          = "ibcwmlhcimymasokhgvn"
-	supabasePublicKey   = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJyb2xlIjoiYW5vbiIsImlhdCI6MTYyOTkzMDc3OCwiZXhwIjoxOTQ1NTA2Nzc4fQ.zcdbd7kDhk7iNSMo8SjsTaXi0wlLNNQcSZkzZ84NUDg"
-	authCallbackAddr    = "localhost:3000"
-)