+++ cmd/insights/contributors.go
@@ -12,10 +12,11 @@ import (
 
 	bubblesTable "github.com/charmbracelet/bubbles/table"
 	"github.com/open-sauced/go-api/client"
+	"github.com/spf13/cobra"
+
 	"github.com/open-sauced/pizza-cli/pkg/api"
 	"github.com/open-sauced/pizza-cli/pkg/constants"
 	"github.com/open-sauced/pizza-cli/pkg/utils"
-	"github.com/spf13/cobra"
 )
 
 type contributorsOptions struct {
@@ -50,7 +51,7 @@ func NewContributorsCommand() *cobra.Command {
 			opts.Repos = append(opts.Repos, args...)
 			return nil
 		},
-		RunE: func(cmd *cobra.Command, args []string) error {
+		RunE: func(cmd *cobra.Command, _ []string) error {
 			endpointURL, _ := cmd.Flags().GetString(constants.FlagNameEndpoint)
 			opts.APIClient = api.NewGoClient(endpointURL)
 			output, _ := cmd.Flags().GetString(constants.FlagNameOutput)
@@ -228,7 +229,7 @@ func findAllContributorsInsights(ctx context.Context, opts *contributorsOptions,
 	waitGroup.Add(1)
 	go func() {
 		defer waitGroup.Done()
-		response, err := findNewRepositoryContributors(ctx, opts.APIClient, repo.Name, opts.Period)
+		response, err := findNewRepositoryContributors(ctx, opts.APIClient, repo.FullName, opts.Period)
 		if err != nil {
 			errorChan <- err
 			return
@@ -240,7 +241,7 @@ func findAllContributorsInsights(ctx context.Context, opts *contributorsOptions,
 	waitGroup.Add(1)
 	go func() {
 		defer waitGroup.Done()
-		response, err := findRecentRepositoryContributors(ctx, opts.APIClient, repo.Name, opts.Period)
+		response, err := findRecentRepositoryContributors(ctx, opts.APIClient, repo.FullName, opts.Period)
 		if err != nil {
 			errorChan <- err
 			return
@@ -252,7 +253,7 @@ func findAllContributorsInsights(ctx context.Context, opts *contributorsOptions,
 	waitGroup.Add(1)
 	go func() {
 		defer waitGroup.Done()
-		response, err := findAlumniRepositoryContributors(ctx, opts.APIClient, repo.Name, opts.Period)
+		response, err := findAlumniRepositoryContributors(ctx, opts.APIClient, repo.FullName, opts.Period)
 		if err != nil {
 			errorChan <- err
 			return
@@ -264,7 +265,7 @@ func findAllContributorsInsights(ctx context.Context, opts *contributorsOptions,
 	waitGroup.Add(1)
 	go func() {
 		defer waitGroup.Done()
-		response, err := findRepeatRepositoryContributors(ctx, opts.APIClient, repo.Name, opts.Period)
+		response, err := findRepeatRepositoryContributors(ctx, opts.APIClient, repo.FullName, opts.Period)
 		if err != nil {
 			errorChan <- err
 			return