+++ cmd/insights/repositories.go
@@ -11,10 +11,11 @@ import (
 
 	bubblesTable "github.com/charmbracelet/bubbles/table"
 	"github.com/open-sauced/go-api/client"
+	"github.com/spf13/cobra"
+
 	"github.com/open-sauced/pizza-cli/pkg/api"
 	"github.com/open-sauced/pizza-cli/pkg/constants"
 	"github.com/open-sauced/pizza-cli/pkg/utils"
-	"github.com/spf13/cobra"
 )
 
 type repositoriesOptions struct {
@@ -51,7 +52,7 @@ func NewRepositoriesCommand() *cobra.Command {
 			opts.Repos = append(opts.Repos, args...)
 			return nil
 		},
-		RunE: func(cmd *cobra.Command, args []string) error {
+		RunE: func(cmd *cobra.Command, _ []string) error {
 			endpointURL, _ := cmd.Flags().GetString(constants.FlagNameEndpoint)
 			opts.APIClient = api.NewGoClient(endpointURL)
 			output, _ := cmd.Flags().GetString(constants.FlagNameOutput)
@@ -197,7 +198,7 @@ func findAllRepositoryInsights(ctx context.Context, opts *repositoriesOptions, r
 	waitGroup.Add(1)
 	go func() {
 		defer waitGroup.Done()
-		response, err := getPullRequestInsights(ctx, opts.APIClient, repo.Name, opts.Period)
+		response, err := getPullRequestInsights(ctx, opts.APIClient, repo.FullName, opts.Period)
 		if err != nil {
 			errorChan <- err
 			return