+++ cmd/root/root.go
@@ -4,14 +4,13 @@ package root
 import (
 	"fmt"
 
+	"github.com/spf13/cobra"
+
 	"github.com/open-sauced/pizza-cli/cmd/auth"
-	"github.com/open-sauced/pizza-cli/cmd/bake"
+	"github.com/open-sauced/pizza-cli/cmd/generate"
 	"github.com/open-sauced/pizza-cli/cmd/insights"
-	repoquery "github.com/open-sauced/pizza-cli/cmd/repo-query"
-	"github.com/open-sauced/pizza-cli/cmd/show"
 	"github.com/open-sauced/pizza-cli/cmd/version"
 	"github.com/open-sauced/pizza-cli/pkg/constants"
-	"github.com/spf13/cobra"
 )
 
 // NewRootCommand bootstraps a new root cobra command for the pizza CLI
@@ -21,7 +20,7 @@ func NewRootCommand() (*cobra.Command, error) {
 		Short: "OpenSauced CLI",
 		Long:  "A command line utility for insights, metrics, and all things OpenSauced",
 		RunE:  run,
-		Args: func(cmd *cobra.Command, args []string) error {
+		Args: func(cmd *cobra.Command, _ []string) error {
 			betaFlag := cmd.Flags().Lookup(constants.FlagNameBeta)
 			if betaFlag.Changed {
 				err := cmd.Flags().Lookup(constants.FlagNameEndpoint).Value.Set(constants.EndpointBeta)
@@ -36,14 +35,24 @@ func NewRootCommand() (*cobra.Command, error) {
 	cmd.PersistentFlags().StringP(constants.FlagNameEndpoint, "e", constants.EndpointProd, "The API endpoint to send requests to")
 	cmd.PersistentFlags().Bool(constants.FlagNameBeta, false, fmt.Sprintf("Shorthand for using the beta OpenSauced API endpoint (\"%s\"). Supersedes the '--%s' flag", constants.EndpointBeta, constants.FlagNameEndpoint))
 	cmd.PersistentFlags().Bool(constants.FlagNameTelemetry, false, "Disable sending telemetry data to OpenSauced")
-	cmd.PersistentFlags().StringP(constants.FlagNameOutput, "o", constants.OutputTable, "The formatting for command output. One of: (table, yaml, csv, json)")
+	cmd.PersistentFlags().StringP("config", "c", "~/.sauced.yaml", "The saucectl config")
+	cmd.PersistentFlags().StringP("log-level", "l", "info", "The logging level. Options: error, warn, info, debug")
+	cmd.PersistentFlags().Bool("tty-disable", false, "Disable log stylization. Suitable for CI/CD and automation")
 
-	cmd.AddCommand(bake.NewBakeCommand())
-	cmd.AddCommand(repoquery.NewRepoQueryCommand())
 	cmd.AddCommand(auth.NewLoginCommand())
+	cmd.AddCommand(generate.NewGenerateCommand())
 	cmd.AddCommand(insights.NewInsightsCommand())
 	cmd.AddCommand(version.NewVersionCommand())
-	cmd.AddCommand(show.NewShowCommand())
+
+	err := cmd.Flags().MarkHidden(constants.FlagNameEndpoint)
+	if err != nil {
+		return nil, err
+	}
+
+	err = cmd.Flags().MarkHidden(constants.FlagNameBeta)
+	if err != nil {
+		return nil, err
+	}
 
 	return cmd, nil
 }