+++ pkg/logging/constants.go
@@ -1,0 +1,8 @@
+package logging
+
+var (
+	LogError = 1
+	LogWarn  = 2
+	LogInfo  = 3
+	LogDebug = 4
+)